+++ pyxel/cli.py
@@ -5,6 +5,7 @@ import re
 import runpy
 import shutil
 import sys
+import tempfile
 import urllib.request
 import zipfile
 
@@ -91,7 +92,7 @@ def _check_dir_exists(dirname):
 
 
 def _make_app_dir():
-    play_dir = os.path.expanduser(os.path.join(pyxel.PYXEL_WORKING_DIR, "play"))
+    play_dir = os.path.join(tempfile.gettempdir(), pyxel.PYXEL_WORKING_DIR, "play")
     pathlib.Path(play_dir).mkdir(parents=True, exist_ok=True)
     for path in glob.glob(os.path.join(play_dir, "*")):
         pid = int(os.path.basename(path))