+++ pyxel/editor/music_editor.py
@@ -31,6 +31,7 @@ class MusicEditor(EditorBase):
             get_field=self.get_field,
             add_pre_history=self.add_pre_history,
             add_post_history=self.add_post_history,
+            cross_filed_copying=True,
         )
 
         # Initialize music picker
@@ -92,7 +93,7 @@ class MusicEditor(EditorBase):
         if index >= pyxel.NUM_CHANNELS:
             return
         music = pyxel.music(self.music_no_var)
-        return music.sequences[index]
+        return music.sounds[index]
 
     def add_pre_history(self, x, y):
         self._history_data = data = {}
@@ -117,7 +118,7 @@ class MusicEditor(EditorBase):
         if is_partial:
             for i in range(self.field_cursor.x):
                 music = pyxel.music(self.music_no_var)
-                sound = pyxel.sound(music.sequences[self.field_cursor.y][i])
+                sound = pyxel.sound(music.sounds[self.field_cursor.y][i])
                 tick += len(sound.notes) * sound.speed
         pyxel.playm(self.music_no_var, tick=tick, loop=self.should_loop_var)
 
@@ -132,13 +133,13 @@ class MusicEditor(EditorBase):
     def __on_undo(self, data):
         self._stop()
         self.music_no_var = data["music_no"]
-        self.field_cursor.move_to(*data["old_cursor_pos"])
+        self.field_cursor.move_to(*data["old_cursor_pos"], False)
         self.field_cursor.field.from_list(data["old_field"])
 
     def __on_redo(self, data):
         self._stop()
         self.music_no_var = data["music_no"]
-        self.field_cursor.move_to(*data["new_cursor_pos"])
+        self.field_cursor.move_to(*data["new_cursor_pos"], False)
         self.field_cursor.field.from_list(data["new_field"])
 
     def __on_hide(self):